requirements.txt
opencv-python
PyYAML
ultralytics
numpy
gpiozero
websockets
aiohttp
